tests/requirements.txt
pytest
requests
python-dotenv
psycopg2-binary